Moved @tryghost/staff-service to Ghost core
fix https://linear.app/ghost/issue/ENG-2067/staff-service

- this commit migrates the code and test for the staff service
  into Ghost core so we can keep all related code together and reduce
  overhead
- as a result of this change, we can remove the package and associated
  references
