+++ ghost/core/core/server/services/staff/email-templates/donation.txt.js
