+++ ghost/core/core/server/services/staff/email-templates/new-free-signup.hbs
