+++ ghost/core/core/server/services/staff/email-templates/new-milestone-received.hbs
