+++ ghost/core/core/server/services/staff/email-templates/new-paid-cancellation.txt.js
