+++ ghost/core/core/server/services/staff/email-templates/partials/styles.hbs
