+++ ghost/core/core/server/services/staff/email-templates/recommendation-received.txt.js
