+++ ghost/core/package.json
@@ -40,7 +40,7 @@
     "test:browser:setup": "npx playwright install",
     "test:ci:e2e": "c8 -c ./.c8rc.e2e.json -o coverage-e2e yarn test:e2e -b --retries=2 --reporter=./test/utils/mocha-retry-reporter.js",
     "test:ci:regression": "yarn test:regression -b --retries=2 --reporter=./test/utils/mocha-retry-reporter.js",
-    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 57 --functions 47 --branches 77 --statements 57 yarn test:integration -b --retries=2 --reporter=./test/utils/mocha-retry-reporter.js",
+    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 57 --functions 47 --branches 76 --statements 57 yarn test:integration -b --retries=2 --reporter=./test/utils/mocha-retry-reporter.js",
     "test:unit:slow": "yarn test:unit --reporter=mocha-slow-test-reporter",
     "test:int:slow": "yarn test:integration --reporter=mocha-slow-test-reporter",
     "test:e2e:slow": "yarn test:e2e --reporter=mocha-slow-test-reporter",
@@ -84,6 +84,7 @@
     "@tryghost/debug": "0.1.32",
     "@tryghost/domain-events": "0.0.0",
     "@tryghost/donations": "0.0.0",
+    "@tryghost/email-addresses": "0.0.0",
     "@tryghost/email-analytics-provider-mailgun": "0.0.0",
     "@tryghost/email-analytics-service": "0.0.0",
     "@tryghost/email-content-generator": "0.0.0",
@@ -150,7 +151,6 @@
     "@tryghost/settings-path-manager": "0.0.0",
     "@tryghost/slack-notifications": "0.0.0",
     "@tryghost/social-urls": "0.1.43",
-    "@tryghost/staff-service": "0.0.0",
     "@tryghost/stats-service": "0.0.0",
     "@tryghost/string": "0.2.12",
     "@tryghost/tiers": "0.0.0",
@@ -189,6 +189,7 @@
     "glob": "8.1.0",
     "got": "11.8.6",
     "gscan": "4.47.0",
+    "handlebars": "4.7.8",
     "human-number": "2.0.4",
     "iconv-lite": "0.6.3",
     "image-size": "1.2.0",
@@ -198,6 +199,7 @@
     "json-stable-stringify": "1.2.0",
     "jsonpath": "1.1.1",
     "jsonwebtoken": "8.5.1",
+    "juice": "9.1.0",
     "keypair": "1.0.4",
     "knex": "2.4.2",
     "knex-migrator": "5.3.0",
