--- ghost/staff-service/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@tryghost/staff-service",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/staff-service",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/email-addresses": "0.0.0",
-    "handlebars": "4.7.8",
-    "juice": "9.1.0",
-    "lodash": "4.17.21",
-    "moment": "2.29.1"
-  }
-}
